Phase 3: Extract ChatUpdate component
- Created ChatUpdate.hpp with public API for update management
- Created ChatUpdate.cpp with ~540 lines of implementation:
  - UpdateState struct and update state management
  - Manifest download and parsing
  - Version comparison (SemanticVersion struct)
  - Package download and installation
  - Backup/restore on installation failure
  - RPC handlers: chatCheckForUpdates, chatInstallUpdate, chatGetUpdateStatus
- Removed update management code from SessionChat.cpp
- Updated CMakeLists.txt to build ChatUpdate.cpp
- Build verified successful on macOS

Update functionality unchanged, including backup/restore on failure.
