+++ src/cpp/session/CMakeLists.txt
@@ -184,6 +184,7 @@ set(SESSION_SOURCE_FILES
    modules/SessionBreakpoints.cpp
    modules/chat/SessionChat.cpp
    modules/chat/ChatInstallation.cpp
+   modules/chat/ChatUpdate.cpp
    modules/SessionCodeSearch.cpp
    modules/SessionConfigFile.cpp
    modules/SessionCRANMirrors.cpp
